new indicator to compute time series in larger timeframe

+++ src/main/java/data/MarketData.java
@@ -68,7 +68,7 @@ public class MarketData {
 		return this.Close;
 	}
 
-	public double getVolume() {
+	public int getVolume() {
 		return this.volume;
 	}
